interface/requirements.txt
streamlit~=1.44.1
numpy~=2.2.4
pillow~=11.1.0